Move new test to r82 package
- #24133
