+++ subprojects/ide/src/crossVersionTest/groovy/org/gradle/plugins/ide/tooling/r82/FetchTasksAction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.plugins.ide.tooling.r81;
+package org.gradle.plugins.ide.tooling.r82;
 
 import org.gradle.tooling.BuildAction;
 import org.gradle.tooling.BuildController;
